+++ core/vil/file_formats/vil_dicom_header.h
@@ -373,8 +373,8 @@ struct vil2_dicom_header_info
 const short VIL2_DICOM_HEADER_UNSPECIFIED = -1;
 const unsigned short VIL2_DICOM_HEADER_UNSPECIFIED_UNSIGNED = (unsigned short) -1;
 const double VIL2_DICOM_HEADER_DEFAULTSIZE = 1.0;
-const double VIL2_DICOM_HEADER_DEFAULTINTERCEPT = 0.0;
-const double VIL2_DICOM_HEADER_DEFAULTSLOPE = 1.0;
+const float VIL2_DICOM_HEADER_DEFAULTINTERCEPT = 0.0;
+const float VIL2_DICOM_HEADER_DEFAULTSLOPE = 1.0;
 
 //: A class to read and write the header part of a dicom file
 //
